refactor(toolbar): extract button sections into dedicated components
- Extract GitHubStatsToolbarButton with useImperativeHandle for overflow coordination
- Extract NotificationCenterToolbarButton with own store subscriptions
- Extract ToolbarLauncherButton for terminal/browser/panel-palette buttons
- Extract ToolbarSettingsButton with context menu and keybinding
- Extract ToolbarProblemsButton with diagnostics shortcut
- Extract ToolbarPortalButton with portal store subscription
- Reduce Toolbar.tsx from 1532 to 1000 lines
- Reduce buttonRegistry useMemo deps from 52 to 22 items
- Update source-level tests to read from new component files
