+++ src/components/Layout/__tests__/Toolbar.shortcuts.test.ts
@@ -3,12 +3,26 @@ import fs from "fs/promises";
 import path from "path";
 
 const TOOLBAR_PATH = path.resolve(__dirname, "../Toolbar.tsx");
+const PROBLEMS_BUTTON_PATH = path.resolve(__dirname, "../ToolbarProblemsButton.tsx");
+const PORTAL_BUTTON_PATH = path.resolve(__dirname, "../ToolbarPortalButton.tsx");
+const SETTINGS_BUTTON_PATH = path.resolve(__dirname, "../ToolbarSettingsButton.tsx");
+const LAUNCHER_BUTTON_PATH = path.resolve(__dirname, "../ToolbarLauncherButton.tsx");
 
 describe("Toolbar shortcut tooltips — issue #3443", () => {
   let source: string;
+  let problemsSource: string;
+  let portalSource: string;
+  let settingsSource: string;
+  let launcherSource: string;
 
   beforeEach(async () => {
-    source = await fs.readFile(TOOLBAR_PATH, "utf-8");
+    [source, problemsSource, portalSource, settingsSource, launcherSource] = await Promise.all([
+      fs.readFile(TOOLBAR_PATH, "utf-8"),
+      fs.readFile(PROBLEMS_BUTTON_PATH, "utf-8"),
+      fs.readFile(PORTAL_BUTTON_PATH, "utf-8"),
+      fs.readFile(SETTINGS_BUTTON_PATH, "utf-8"),
+      fs.readFile(LAUNCHER_BUTTON_PATH, "utf-8"),
+    ]);
   });
 
   describe("useKeybindingDisplay hooks", () => {
@@ -17,11 +31,11 @@ describe("Toolbar shortcut tooltips — issue #3443", () => {
     });
 
     it("uses dynamic hook for panel.toggleDiagnostics", () => {
-      expect(source).toContain('useKeybindingDisplay("panel.toggleDiagnostics")');
+      expect(problemsSource).toContain('useKeybindingDisplay("panel.toggleDiagnostics")');
     });
 
     it("uses dynamic hook for panel.togglePortal", () => {
-      expect(source).toContain('useKeybindingDisplay("panel.togglePortal")');
+      expect(portalSource).toContain('useKeybindingDisplay("panel.togglePortal")');
     });
 
     it("uses dynamic hook for notes.openPalette", () => {
@@ -29,7 +43,7 @@ describe("Toolbar shortcut tooltips — issue #3443", () => {
     });
 
     it("uses dynamic hook for app.settings", () => {
-      expect(source).toContain('useKeybindingDisplay("app.settings")');
+      expect(settingsSource).toContain('useKeybindingDisplay("app.settings")');
     });
   });
 
@@ -45,21 +59,26 @@ describe("Toolbar shortcut tooltips — issue #3443", () => {
 
   describe("no manual ternary tooltip patterns", () => {
     it("does not use manual ternary for terminal shortcut", () => {
+      // Terminal shortcut is now in the launcher button component
       expect(source).not.toContain("terminalShortcut ?");
+      expect(launcherSource).not.toContain("terminalShortcut ?");
     });
 
     it("does not use manual ternary for browser shortcut", () => {
       expect(source).not.toContain("browserShortcut ?");
+      expect(launcherSource).not.toContain("browserShortcut ?");
     });
   });
 
   describe("createTooltipWithShortcut usage", () => {
     it("uses createTooltipWithShortcut for terminal tooltip", () => {
-      expect(source).toContain('createTooltipWithShortcut("Open Terminal", terminalShortcut)');
+      expect(launcherSource).toContain('tooltipLabel: "Open Terminal"');
+      expect(launcherSource).toContain("createTooltipWithShortcut(config.tooltipLabel, shortcut)");
     });
 
     it("uses createTooltipWithShortcut for browser tooltip", () => {
-      expect(source).toContain('createTooltipWithShortcut("Open Browser", browserShortcut)');
+      expect(launcherSource).toContain('tooltipLabel: "Open Browser"');
+      expect(launcherSource).toContain("createTooltipWithShortcut(config.tooltipLabel, shortcut)");
     });
 
     it("uses createTooltipWithShortcut for notes tooltip", () => {
@@ -67,20 +86,20 @@ describe("Toolbar shortcut tooltips — issue #3443", () => {
     });
 
     it("uses createTooltipWithShortcut for settings tooltip", () => {
-      expect(source).toContain('createTooltipWithShortcut("Open Settings", settingsShortcut)');
+      expect(settingsSource).toContain(
+        'createTooltipWithShortcut("Open Settings", settingsShortcut)'
+      );
     });
 
     it("uses createTooltipWithShortcut for problems tooltip with dynamic shortcut", () => {
-      expect(source).toContain(
+      expect(problemsSource).toContain(
         'createTooltipWithShortcut("Show Problems Panel", diagnosticsShortcut)'
       );
     });
 
     it("uses createTooltipWithShortcut for portal tooltip", () => {
-      const portalBlock = source.match(/"portal-toggle":\s*\{[\s\S]*?isAvailable/);
-      expect(portalBlock).not.toBeNull();
-      expect(portalBlock![0]).toContain("createTooltipWithShortcut");
-      expect(portalBlock![0]).toContain("portalShortcut");
+      expect(portalSource).toContain("createTooltipWithShortcut");
+      expect(portalSource).toContain("portalShortcut");
     });
 
     it("uses createTooltipWithShortcut for sidebar tooltip with dynamic shortcut", () => {
@@ -99,20 +118,6 @@ describe("Toolbar shortcut tooltips — issue #3443", () => {
       expect(deps).toContain("sidebarShortcut");
     });
 
-    it("includes diagnosticsShortcut in useMemo deps", () => {
-      const depsMatch = source.match(/\}\),\s*\[([^\]]+)\]\s*\);/s);
-      expect(depsMatch).not.toBeNull();
-      const deps = depsMatch![1];
-      expect(deps).toContain("diagnosticsShortcut");
-    });
-
-    it("includes portalShortcut in useMemo deps", () => {
-      const depsMatch = source.match(/\}\),\s*\[([^\]]+)\]\s*\);/s);
-      expect(depsMatch).not.toBeNull();
-      const deps = depsMatch![1];
-      expect(deps).toContain("portalShortcut");
-    });
-
     it("includes notesShortcut in useMemo deps", () => {
       const depsMatch = source.match(/\}\),\s*\[([^\]]+)\]\s*\);/s);
       expect(depsMatch).not.toBeNull();
@@ -120,11 +125,16 @@ describe("Toolbar shortcut tooltips — issue #3443", () => {
       expect(deps).toContain("notesShortcut");
     });
 
-    it("includes settingsShortcut in useMemo deps", () => {
-      const depsMatch = source.match(/\}\),\s*\[([^\]]+)\]\s*\);/s);
-      expect(depsMatch).not.toBeNull();
-      const deps = depsMatch![1];
-      expect(deps).toContain("settingsShortcut");
+    it("diagnosticsShortcut is in ToolbarProblemsButton", () => {
+      expect(problemsSource).toContain('useKeybindingDisplay("panel.toggleDiagnostics")');
+    });
+
+    it("portalShortcut is in ToolbarPortalButton", () => {
+      expect(portalSource).toContain('useKeybindingDisplay("panel.togglePortal")');
+    });
+
+    it("settingsShortcut is in ToolbarSettingsButton", () => {
+      expect(settingsSource).toContain('useKeybindingDisplay("app.settings")');
     });
   });
 });
